# openapi.yaml
openapi: 3.0.3
info:
  title: Secure Legal DMS API
  version: 0.1.0
  description: >
    Backend contract for the SIH 2026 secure document management system.
    This file is the source of truth. Do not change an endpoint without
    telling the whole team in writing.
servers:
  - url: http://localhost:8000/api/v1
    description: Local development
...
security:
  - sessionToken: []
paths:
  /auth/login:
    post:
      tags: [auth]
      summary: Step 1 of login. Returns an MFA challenge, not a session.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [service_number, password]
              properties:
                service_number:
                  type: string
                  example: DL-SI-40219
                password:
                  type: string
                  format: password
      responses:
        '200':
          description: Credentials accepted, MFA code required
          content:
            application/json:
              schema:
                type: object
                properties:
                  mfa_token:
                    type: string
                    description: Short-lived token to pass to /auth/mfa/verify
                  expires_in:
                    type: integer
                    example: 300
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/mfa/verify:
    post:
      tags: [auth]
      summary: >
        Step 2 of the authenticator-app path. Exchanges a TOTP code for a
        session. The account must have an authenticator enrolled - a
        password alone never yields a session - and the challenge is
        discarded after 5 wrong codes.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [mfa_token, code]
              properties:
                mfa_token:
                  type: string
                code:
                  type: string
                  pattern: '^[0-9]{6}$'
                  example: "418203"
      responses:
        '200':
          description: Session issued
          content:
            application/json:
              schema:
                type: object
                properties:
                  session_token:
                    type: string
                  expires_at:
                    type: string
                    format: date-time
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/otp/request:
    post:
      tags: [auth]
      summary: >
        Sign-in, step 1. Checks service number and password, and that the
        mobile number is the officer's registered one; only then texts a
        6-digit code to it. Codes expire in 5 minutes; at most 3 per
        officer per 15 minutes, 60 s apart. Only an HMAC of the code is
        stored.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [service_number, password, mobile_number]
              properties:
                service_number:
                  type: string
                password:
                  type: string
                mobile_number:
                  type: string
                  description: E.164, or ten digits read as +91
                  example: "9876543210"
      responses:
        '202':
          description: Code sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  otp_token:
                    type: string
                    format: uuid
                    description: >
                      Names the code to verify in step 2. Random per request and
                      issued only to a caller that passed the password, so nobody
                      else can spend this officer's attempts.
                  message:
                    type: string
                  expires_in:
                    type: integer
                  resend_after:
                    type: integer
        '400':
          description: A field is missing, or the mobile number is not valid
        '401':
          description: >
            Invalid credentials - the same answer for a wrong password and
            for a mobile number that is not the officer's
        '409':
          description: >
            no_mobile_registered - the password was right but the account has no
            mobile number, so no code can be sent
        '429':
          description: Too many requests from this address, or codes for this officer
        '502':
          description: The SMS provider could not send the code
  /auth/otp/verify:
    post:
      tags: [auth]
      summary: >
        Sign-in, step 2. Exchanges the code for a session. Keyed on the
        otp_token from step 1, so only the caller that passed the password
        can spend the code's attempts. Each guess costs one of 5; a code
        works once, and requesting a new code retires the previous one.
        Unknown token, wrong, expired, exhausted or used code all return
        the same 401.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [otp_token, otp]
              properties:
                otp_token:
                  type: string
                  format: uuid
                  description: From the step 1 response
                otp:
                  type: string
                  pattern: '^[0-9]{6}$'
      responses:
        '200':
          description: Session issued
          content:
            application/json:
              schema:
                type: object
                properties:
                  session_token:
                    type: string
                  expires_at:
                    type: string
                    format: date-time
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: invalid_code
        '429':
          description: Too many attempts from this address
...
components:
  securitySchemes:
    sessionToken:
      type: http
      scheme: bearer
      description: Opaque session token from /auth/mfa/verify
...
  schemas:
...
    Rank:
      type: string
      enum: [constable, head_constable, sub_inspector, inspector, dcp, prosecutor, judge, forensic_analyst]
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        service_number:
          type: string
        rank:
          $ref: '#/components/schemas/Rank'
        station:
          type: string
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: forbidden
        message:
          type: string
          example: You are not assigned to this case.
  responses:
    Unauthorized:
      description: Not logged in, or session expired
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...